// src/type.tsx
export interface Data {
    id: string;
    FECHA: string;
    numero: number;
    descripcion: string;
}
export interface Datad {
    IDDEBE: string;
    CUENTAD: string;
    CANTIDADD: number;
    RESULTADOD: String;
    TIPOD: String;

}
export interface Datah {
    IDHABER: string;
    CUENTAH: string;
    CANTIDADH: number;
    RESULTADOH: string;
    TIPOH: string;
}